basic abstraction of functions
